+++ lib/ask-tools-shell.rb
@@ -1,13 +1,4 @@
 # frozen_string_literal: true
 
 require "ask-tools"
-
-require_relative "ask/tools/shell/version"
-require_relative "ask/tools/shell/bash"
-require_relative "ask/tools/shell/read"
-require_relative "ask/tools/shell/write"
-require_relative "ask/tools/shell/edit"
-require_relative "ask/tools/shell/glob"
-require_relative "ask/tools/shell/grep"
-require_relative "ask/tools/shell/code"
-require_relative "ask/tools/shell"
+require "ask/tools/shell"